move code back for review
